Also applying same check for wordpress (#542)

+++ src/platforms/wordpress/Framework/Configurations.php
@@ -115,7 +115,9 @@ class Configurations extends BaseConfigurations
             throw new \RuntimeException('Outline not found', 404);
         }
 
-        Folder::delete($path);
+        if (file_exists($path)) {
+            Folder::delete($path);
+        }
     }
 
     /**
